pyproject.toml
[project]
name = "atsphinx-github-io"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "Kazuya Takei", email = "myself@attakei.net" }
]
dependencies = [
    "atsphinx-color-text~=0.1.0",
    "sphinx~=7.2.0",
    "pydata-sphinx-theme~=0.13.0",
    "sphinx_design~=0.5.0",
    "python-dotenv~=1.0.1",
    "atsphinx-mini18n~=0.3.0",
    "atsphinx-footnotes>=0.2.0",
]
license = {text = "Apache-2.0"}
readme = "README.rst"
requires-python = ">=3.12"

[tool.uv]
dev-dependencies = [
    "sphinx-autobuild>=2021.3.14",
    "sphinx-intl~=2.1.0",
    "pygithub~=2.3.0",
]

[tool.ruff.lint]
select = ["C90", "D", "E", "F", "I", "W"]

[tool.ruff.lint.pydocstyle]
convention = "pep257"
